Viet lt/6.11.6/crud prescription (#35)
* HauTC to dev (#31) (#33)

* #6.2.1.1 - HauTC Invoice DisplayInvoice api

* Dev to HauTC (#29)

* test

* prescription-be

* CRUDEmployee

* up MedicineInformationService

* up MedicineController

* CRUDCustomer

* crudCustomer

* up

* 99.3 ThangNV Medicine-detail Create-api

# Conflicts:
#	A0223I1_Pharmacy_Professional_BE/build.gradle
#	A0223I1_Pharmacy_Professional_BE/src/main/java/com/codegym/a0223i1_pharmacy_professional_be/controller/informationmanagement/MedicineInformationManagementController.java
#	A0223I1_Pharmacy_Professional_BE/src/main/resources/application.properties

* update full be

* #6.2.1.1 - HauTC Invoice DisplayInvoice api (#28)

---------











* #2 - HauTC Update Entity

---------


* okay

* Viet BE

---------
