+++ A0223I1_Pharmacy_Professional_BE/src/main/java/com/codegym/a0223i1_pharmacy_professional_be/controller/informationmanagement/PrescriptionManagementController.java
@@ -1,8 +1,10 @@
 package com.codegym.a0223i1_pharmacy_professional_be.controller.informationmanagement;
 
+import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
+@CrossOrigin("*")
 public class PrescriptionManagementController {
     // QUản lý thông tin toa thuốc
 }
