+++ A0223I1_Pharmacy_Professional_BE/src/main/java/com/codegym/a0223i1_pharmacy_professional_be/entity/AccountRole.java
@@ -26,6 +26,7 @@ public class AccountRole {
             CascadeType.PERSIST, CascadeType.MERGE,
             CascadeType.DETACH, CascadeType.REFRESH
     })
+
     @JoinColumn(name = "account_id", referencedColumnName = "account_id")
     @JsonBackReference
     private Account account;
