+++ A0223I1_Pharmacy_Professional_BE/src/main/java/com/codegym/a0223i1_pharmacy_professional_be/entity/Medicine.java
@@ -73,7 +73,4 @@ public class Medicine {
     @JsonBackReference
     List<InvoiceDetail> invoiceDetails;
 
-    @OneToMany(mappedBy = "medicine")
-    @JsonBackReference
-    List<PrescriptionDetail> prescriptionDetails;
 }
