+++ A0223I1_Pharmacy_Professional_BE/src/main/java/com/codegym/a0223i1_pharmacy_professional_be/entity/Prescription.java
@@ -35,6 +35,9 @@ public class Prescription {
 
     private Boolean deleteFlag;
 
+    @Transient
+    private String symptomName;
+
     @OneToMany(mappedBy = "prescription")
     @JsonBackReference
     List<PrescriptionDetail> prescriptionDetails;
