+++ A0223I1_Pharmacy_Professional_BE/src/main/java/com/codegym/a0223i1_pharmacy_professional_be/entity/PrescriptionDetail.java
@@ -19,13 +19,6 @@ public class PrescriptionDetail {
             CascadeType.PERSIST, CascadeType.MERGE,
             CascadeType.DETACH, CascadeType.REFRESH
     })
-    @JoinColumn(name = "medicine_id", referencedColumnName ="medicine_id")
-    private Medicine medicine;
-
-    @ManyToOne(cascade = {
-            CascadeType.PERSIST, CascadeType.MERGE,
-            CascadeType.DETACH, CascadeType.REFRESH
-    })
     @JoinColumn(name = "prescription_id", referencedColumnName ="prescription_id")
     @JsonBackReference
     private Prescription prescription;
