+++ A0223I1_Pharmacy_Professional_BE/src/main/java/com/codegym/a0223i1_pharmacy_professional_be/entity/Symptom.java
@@ -3,6 +3,8 @@ package com.codegym.a0223i1_pharmacy_professional_be.entity;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.util.Set;
+
 @Getter
 @Setter
 @NoArgsConstructor
@@ -16,4 +18,7 @@ public class Symptom {
     private Integer symptomId;
 
     private String symptomName;
+
+    @OneToMany(mappedBy = "symptom", cascade = CascadeType.ALL)
+    private Set<Prescription> prescriptions;
 }
