+++ A0223I1_Pharmacy_Professional_BE/src/main/java/com/codegym/a0223i1_pharmacy_professional_be/repository/informationmanagement/MedicineRepository.java
@@ -16,10 +16,11 @@ import java.util.Optional;
 @Repository
 @Transactional
 public interface MedicineRepository extends JpaRepository<Medicine, String> {
-
-
     @Query("SELECT m FROM Medicine m JOIN FETCH m.medicineImgs WHERE m.medicineId = :id")
     Medicine findMedicineByMedicineId(String id);
+
+    @Query(value = "select id, name from medicine where id= :id", nativeQuery = true)
+    Optional<Medicine> findById(@Param("id") String id);
 
 }
