+++ A0223I1_Pharmacy_Professional_BE/src/main/java/com/codegym/a0223i1_pharmacy_professional_be/repository/informationmanagement/PrescriptionRepository.java
@@ -1,4 +1,5 @@
-package com.codegym.a0223i1_pharmacy_professional_be.repository.prescriptionmanagement;
+package com.codegym.a0223i1_pharmacy_professional_be.repository.informationmanagement;
+
 
 import com.codegym.a0223i1_pharmacy_professional_be.entity.Prescription;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -9,9 +10,9 @@ import org.springframework.data.repository.query.Param;
 import java.util.List;
 
 public interface PrescriptionRepository extends JpaRepository<Prescription, String> {
-
-    @Query(value = "SELECT id, name, target, note, treatment_period, delete_flag, symptom_id  " +
+    @Query(value = "SELECT p.id, p.name, p.target, p.note, p.treatment_period, p.delete_flag, symptom_id, s.name AS symptomName " +
             "FROM prescription p " +
+            "JOIN symptom s ON p.symptom_id = s.id " +
             "WHERE p.delete_flag = 1", nativeQuery = true)
     List<Prescription> findAllPrescription();
 
@@ -24,6 +25,7 @@ public interface PrescriptionRepository extends JpaRepository<Prescription, Stri
             "JOIN detail_prescription dp ON dp.pres_id = p.id " +
             "WHERE dp.id = :id", nativeQuery = true)
     Prescription findPrescriptionByDetailId(@Param("id") Long id);
+
 
 
 }
