+++ A0223I1_Pharmacy_Professional_BE/src/main/java/com/codegym/a0223i1_pharmacy_professional_be/repository/informationmanagement/SymptomRepository.java
@@ -1,4 +1,4 @@
-package com.codegym.a0223i1_pharmacy_professional_be.repository.prescriptionmanagement;
+package com.codegym.a0223i1_pharmacy_professional_be.repository.informationmanagement;
 
 import com.codegym.a0223i1_pharmacy_professional_be.entity.Symptom;
 import org.springframework.data.jpa.repository.JpaRepository;
