+++ A0223I1_Pharmacy_Professional_BE/src/main/java/com/codegym/a0223i1_pharmacy_professional_be/service/implement/informationmanagement/SymptomServiceImpl.java
@@ -1,13 +1,13 @@
-package com.codegym.a0223i1_pharmacy_professional_be.service.interfaceservice.prescriptionmanagement;
-
+package com.codegym.a0223i1_pharmacy_professional_be.service.implement.informationmanagement;
 
 import com.codegym.a0223i1_pharmacy_professional_be.entity.Symptom;
-import com.codegym.a0223i1_pharmacy_professional_be.repository.prescriptionmanagement.SymptomRepository;
-import com.codegym.a0223i1_pharmacy_professional_be.service.implement.prescriptionmanagement.SymptomService;
+import com.codegym.a0223i1_pharmacy_professional_be.repository.informationmanagement.SymptomRepository;
+import com.codegym.a0223i1_pharmacy_professional_be.service.interfaceservice.informationmanagement.SymptomService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+
 @Service
 public class SymptomServiceImpl implements SymptomService {
     private SymptomRepository symptomRepository;
@@ -23,14 +23,13 @@ public class SymptomServiceImpl implements SymptomService {
     }
 
     @Override
-    public Symptom addSymptom(Symptom symptom) {
-        return symptomRepository.save(symptom);
-    }
-
-    @Override
     public Symptom findSymptomByPrescriptionId(String id) {
         return symptomRepository.findSymptomByPrescriptionId(id);
     }
 
 
+    @Override
+    public Symptom findById(Integer id) {
+        return symptomRepository.findById(id).orElse(null);
+    }
 }
