--- A0223I1_Pharmacy_Professional_BE/src/main/java/com/codegym/a0223i1_pharmacy_professional_be/service/implement/prescriptionmanagement/PrescriptionService.java
@@ -1,23 +1,0 @@
-package com.codegym.a0223i1_pharmacy_professional_be.service.implement.prescriptionmanagement;
-
-
-import com.codegym.a0223i1_pharmacy_professional_be.entity.Prescription;
-import com.codegym.a0223i1_pharmacy_professional_be.entity.Symptom;
-
-import java.util.List;
-
-public interface PrescriptionService {
-    List<Prescription> findAll();
-
-    Prescription addPrescription(Prescription prescription, Symptom symptom);
-
-    Prescription addPrescription(Prescription prescription);
-
-    String generateNextCode();
-
-    Prescription updatePrescription(Prescription prescription, Symptom symptom);
-
-    Prescription findPrescriptionById(String id);
-    void deletePrescription(String id);
-    Prescription findPrescriptionByDetailId(Long id);
-}
