+++ A0223I1_Pharmacy_Professional_BE/src/main/java/com/codegym/a0223i1_pharmacy_professional_be/service/interfaceservice/informationmanagement/MedicineInformationService.java
@@ -3,9 +3,12 @@ package com.codegym.a0223i1_pharmacy_professional_be.service.interfaceservice.in
 import com.codegym.a0223i1_pharmacy_professional_be.entity.Medicine;
 
 import java.util.List;
+import java.util.Optional;
 
 public interface MedicineInformationService {
     List<Medicine> getAllMedicine();
     Medicine findMedicineById(String id);
     void deleteMedicine(Medicine medicine);
+
+    Optional<Medicine> findById(String id);
 }
