+++ A0223I1_Pharmacy_Professional_BE/src/main/java/com/codegym/a0223i1_pharmacy_professional_be/service/interfaceservice/informationmanagement/SymptomService.java
@@ -1,5 +1,4 @@
-package com.codegym.a0223i1_pharmacy_professional_be.service.implement.prescriptionmanagement;
-
+package com.codegym.a0223i1_pharmacy_professional_be.service.interfaceservice.informationmanagement;
 
 import com.codegym.a0223i1_pharmacy_professional_be.entity.Symptom;
 
@@ -7,6 +6,6 @@ import java.util.List;
 
 public interface SymptomService {
     List<Symptom> findAll();
-    Symptom addSymptom(Symptom symptom);
     Symptom findSymptomByPrescriptionId(String id);
+    Symptom findById (Integer id);
 }
